Application Services and API Controllers

+++ src/MealPlannerAPI.Domain/Recipes/RecipeIngredient.cs
@@ -7,10 +7,10 @@ namespace MealPlannerAPI.Recipes
     {
         public Guid RecipeId { get; set; }
         public string Name { get; set; } = null!;
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string Unit { get; set; } = null!;
         protected RecipeIngredient() { }
-        public RecipeIngredient(Guid id, Guid recipeId, string name, double quantity, string unit)
+        public RecipeIngredient(Guid id, Guid recipeId, string name, decimal quantity, string unit)
             : base(id)
         {
             RecipeId = recipeId;
